[new feature]: email ID of subscribers can be listed in a plain text file & gitblag does what it has to do :)
How does this work?

List the email ID of all the subscribers, an email ID per line, in a
text file.

Next, go to the remote bare repository & give git the path to this
text file:

    git config --local --add hooks.recipientlist /path/to/recipient-list.txt

BANG! From the next time on, whenever a NEW POST is committed & pushed
to the bare repository, an email about the blag post is sent to each
email ID in the `recipient-list.txt' file.